Protected member accessed from outside the class
16        bot["plugins"]["slack"].routers["event"]._routes["team_join"]["*"]["*"][0][0]

Accessing a protected member (a member prefixed with _) of a class from outside that class is not recommended, since the creator of that class did not intend this member to be exposed. If accesing this attribute outside of the class is absolutely needed, refactor it such that it becomes part of the public interface of the class.